Make dfdx optional via feature
