+++ gymnasium/Cargo.toml
@@ -12,15 +12,19 @@ rust-version.workspace = true
 version.workspace = true
 
 [dependencies]
-dfdx = { workspace = true }
+dfdx = { workspace = true, optional = true }
 gymnasium_sys = { workspace = true }
+itertools = { workspace = true }
+ndarray = { workspace = true }
 num-traits = { workspace = true }
+numpy = { workspace = true }
 pyo3 = { workspace = true }
 rand = { workspace = true }
 thiserror = { workspace = true }
-numpy = { workspace = true }
-ndarray = { workspace = true }
 
+[features]
+default = []
+dfdx = ["dep:dfdx"]
 
 [lib]
 name = "gymnasium"
