+++ gymnasium/src/lib.rs
@@ -3,16 +3,18 @@
 // pub mod env;
 // pub mod registry;
 pub mod space;
+#[cfg(feature = "dfdx")]
+pub mod space_dfdx;
 pub mod utils;
 
 /// Prelude module for the gymnasium crate that re-exports the most commonly used items.
 pub mod prelude {
     // pub use crate::env::Env;
     // pub use crate::registry::Registry;
-    pub use crate::space::{
-        BoxSpace, BoxSpaceIdentical, BoxSpaceIndependent, DiscreteSpace, MultiBinarySpace,
-        MultiDiscreteSpace, Space, TextSpace,
-    };
+    // pub use crate::space::{
+    //     BoxSpace, BoxSpaceIdentical, BoxSpaceIndependent, DiscreteSpace, MultiBinarySpace,
+    //     MultiDiscreteSpace, Space, TextSpace,
+    // };
     pub use crate::{GymnasiumError, GymnasiumResult};
 }
