+++ gymnasium/src/space_dfdx/fundamental/multi_binary.rs
