+++ gymnasium/src/space_dfdx/fundamental/text.rs
